Let measure_time() output also <file>:<lineno>
The lazy way to measure time is now

	measure_time("");

since that will output

	<file>:<lineno>: : <time>

for every such call (providing enough information to pinpoint performance
bottlenecks).

Since it requires the __VA_ARGS__ feature which this developer suspects to
be unsupported by Visual C++ versions currently in use, this functionality
is guarded by an '#ifdef __GNUCC__'.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>